moved support files to cypress directory, fixed header.spec.tsx
