+++ .gitignore
@@ -10,7 +10,9 @@
 
 # testing
 /coverage
-/cypress
+/cypress/downloads
+/cypress/screenshots
+
 
 # production
 /dist
