+++ cypress/support/component-index.html
