+++ cypress/support/component.ts
