+++ cypress/support/e2e.ts
@@ -14,4 +14,4 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
-import './commands'
+import '../../tests/e2e/support/commands'
